Redefines JSONObject alias
Rather than pointing to the `NSDictionary` equivalent of `[AnyHashable: Any]` it now points to the more JSON-y `[String: Any]`.

I originally avoided this because it can be uncomfortable to cast to when dealing directly with `JSONSerialization`, which returns an `Any` which is an `NSDictionary` half the time. But because Medea only deals with JSON arrays and JSON objects in its interface, I realized there was no reason not to do it (and we gain some correctness to boot).